refactor: organize css & animate thumbnails
I created multiple files for the css code so that it is easier to maintain and create new pages. This way the site can be expanded a lot easier without breaking things. I also made it so that the thumbnails slightly increase in size when the mouse is hovered on them. I also made it so that if you click on the thumbnail of the procedural planets project, it opens a new web page where you can test out the demo application. I might implement this functionality into more thumbnails.